requirements.txt
Django==1.10.5
PyMySQL==0.7.9

# use latest markdown2 to get correct task-list plugin
-e git+https://github.com/trentm/python-markdown2.git#egg=markdown2

qiniu
djangorestframework
